+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/block/direction/DirectionType.java
@@ -9,6 +9,8 @@ public interface DirectionType<T extends Enum<T> & StringRepresentable> {
 
     T getDefault();
 
+    Direction extractDirection(T direction);
+
     T getDirection(Direction clickedFace, Direction playerFacing, float playerPitch);
 
     T rotate(T direction);